+++ src/main/kotlin/view/AlgorithmSubMenu.kt
@@ -100,58 +100,3 @@ fun AlgorithmSubMenu(viewModel: CanvasViewModel) {
         }
     }
 }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ShortestPathDialog(
-    showDialog: MutableState<Boolean>,
-    onPathSelected: (Int, Int) -> Unit
-) {
-    var startIdx by remember { mutableStateOf(0) }
-    var endIdx by remember { mutableStateOf(0) }
-
-    if (showDialog.value) {
-        AlertDialog(
-            onDismissRequest = { showDialog.value = false },
-            buttons = {
-                Button(onClick = {
-                    onPathSelected(startIdx, endIdx)
-                    showDialog.value = false
-                }) {
-                    Text("Найти кратчайший путь")
-                }
-            },
-            text = {
-                Column {
-                    TextField(
-                        value = startIdx.toString(),
-                        onValueChange = { startIdx = it.toIntOrNull() ?: 0 },
-                        label = { Text("Введите точку отправления") }
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    TextField(
-                        value = endIdx.toString(),
-                        onValueChange = { endIdx = it.toIntOrNull() ?: 0 },
-                        label = { Text("Введите точку назначения") }
-                    )
-                }
-            }
-        )
-    }
-}
-
-/*
-        Button(
-            onClick = {
-                val graph = viewModel.graph
-                val painter = GraphPainterByDjikstra(graph, viewModel.graphViewModel, startIdx)
-                painter.paint()
-            },
-            enabled = true,
-            modifier = Modifier.padding(top = 3.dp),
-        ) {
-            Text(
-                text = "Кратчайший путь алгоритмом Дейкстры",
-            )
-        }
- */
